+++ src/main/java/cam72cam/mod/gui_v2/core/actions/IClickable.java
@@ -9,5 +9,5 @@ public interface IClickable {
     /**
      * PRIMARY for left click, otherwise SECONDARY
      */
-    boolean onClick(Player.Hand hand, float x, float y);
+    boolean onClick(Player.Hand hand, int x, int y);
 }
